symcomp-server/requirements.txt
asgiref==3.9.0
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
dj-rest-auth==7.0.1
Django==5.2.4
django-allauth==65.9.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
python-dotenv==1.1.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.10.1
gherkin-official==29.0.0
git-filter-repo==2.47.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
parse==1.20.2
parse_type==0.6.4
Pillow==11.3.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.9.0
pypng==0.20220715.0
pytest==8.4.1
pytest-bdd==8.1.0
pytest-django==4.11.1
PyYAML==6.0.3
qrcode==7.4.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.27.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.11.0
gunicorn==21.2.0
